final/Intelligent_Financial_Data_Integration/scripts/requirements.txt
google-generativeai==0.1.0rc2
typing-extensions>=4.0.0
